requirements.txt
langchain==0.3.2
langchain_core==0.3.9
langchain_groq==0.2.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.37.1
